pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "verify-oidc-token"
version = "0.1.0b1"
description = "A tool to verify OIDC tokens"
requires-python = ">=3.8"
authors = [
  { name = "Andrew Grigorev" , email = "andrew@ei-grad.ru" },
]
urls = { GitHub = "https://github.com/ei-grad/verify-oidc-token" }
dependencies = [
    "requests>=2.25.0,<3.0.0",
    "pyjwt>=2.0.0,<3.0.0",
    "cryptography>=43.0.3",
]

[project.scripts]
verify-oidc-token = "verify_oidc_token.cli:main"

